backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
Pillow==10.2.0
aiogram==3.3.0
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
aiosqlite==0.19.0